LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/report-clips-sample.https-expected.txt


PASS Unsafe eval violation sample is clipped to 40 characters.
FAIL Function constructor - the other kind of eval - is clipped. assert_equals: expected "Function|(a,b) {return '12345678901234567890123" but got "Function|function anonymous(a,b) {return '12345"
PASS Trusted Types violation sample is clipped to 40 characters excluded the sink name.
